FB-245 union service works (#112)
* FB-245 fix models

* FB-245 fix order fields

+++ backend-api/src/main/kotlin/com/rarible/flow/api/controller/NftCollectionApiController.kt
@@ -59,7 +59,7 @@ class NftCollectionApiController(
         } else {
             val last = data.last()
             FlowNftCollectionsDto(
-                data.size,
+                data.size.toLong(),
                 NftCollectionContinuation(last.createdDate, last.id).toString(),
                 data.map { convert(it)!! }
             )
